added cameraIO/ZED, autonomous movement, apriltag folllowing, fixed module initialization
Camera.java currently uses one camera and kinda asssumes its the ZED, the whole structure of the subsystem needs to be modified a bit.  Added position control through a trajectory of position/velocity setpoints, created two commands: move to a given pose, and follow an apriltag. The modules were currently initialized with the same ID so I fixed that to increment the motor ID with the module index